Relationships.

+++ src/main/java/com/demo/healthcare/demo/JpaRepositoryDemo.java
@@ -27,20 +27,20 @@ public class JpaRepositoryDemo implements CommandLineRunner {
         patientRepository.findAll()
                 .forEach(e -> System.out.println("Patient: " + e.getName()));
 
-        Optional<Patient> patient = patientRepository.findById(Math.toIntExact(Long.valueOf(2)));
+        Optional<Patient> patient = patientRepository.findById(Long.valueOf(2));
         System.out.println("Patient ID 2: " + patient.get().getName());
 
         // Update Data
-        Patient patientIdOne = patientRepository.findById(Math.toIntExact(Long.valueOf(1))).get();
+        Patient patientIdOne = patientRepository.findById(Long.valueOf(1)).get();
         patientIdOne.setName("John Doe Updated");
         patientRepository.save(patientIdOne);
 
         // Delete Data
-        Patient patientIdTwo = patientRepository.findById(Math.toIntExact(Long.valueOf(2))).get();
+        Patient patientIdTwo = patientRepository.findById(Long.valueOf(2)).get();
         patientRepository.delete(patientIdTwo);
 
         // Check if Exists
-        boolean exists = patientRepository.existsById(Math.toIntExact(1L));
+        boolean exists = patientRepository.existsById(1L);
         System.out.println("Patient ID 1 exists: " + exists);
     }
 }
